Testing tests work!
No longer using an environment variable. There are methods that patch and unpatch Wisper, and the Testing module keeps track of whether it's currently patched.

Only the Wisper::Testing spec includes the testing module, and it sets an environment variable so that the patching will not happen at inclusion time. So no other specs should be affected.

Finally, added a spec that covers the unpatching and ensures that when unpatched, the original `broadcast` method is restored.