Add a license file and slim the gem artifact to just lib/bins
Many projects read in the LICENSE file of a project to perform license
acceptance for dependencies. I've added the license file here. Github
uses this as well to aid in search so it's quite handy. I also slimmed
down the gem by only shipping the lib/bin files in the gem. There's no
need for the docs/test/hidden files in a gem artifact that goes deep in
a filesystem.

Signed-off-by: Tim Smith <tsmith@chef.io>